requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
python-dotenv==1.0.1
httpx[http2]==0.26.0
sqlalchemy>=2.0.27,<3.0.0
websockets==12.0
aiosqlite==0.19.0
tenacity==8.2.3
databases[sqlite]==0.9.0
pytest==8.0.0
pytest-asyncio==0.23.3
ecdsa==0.18.0
fastapi-middleware==0.2.1